+++ modules/pilon.py
@@ -15,18 +15,47 @@ def indexSequenceBowtie2(referenceFile, threads):
 
 
 # Mapping with Bowtie2
-def mappingBowtie2(fastq_files, referenceFile, threads, outdir):
+def mapping_bowtie2(fastq_files, reference_file, outdir, keep_bam=False, threads=1):
+    """
+    Map reads against a reference fasta file
+
+    Parameters
+    ----------
+    fastq_files : list
+        List of fastq files
+    reference_file : str
+        Path to the reference file (the assembly)
+    outdir : str
+        Path to the output directory
+    keep_bam : bool, default False
+        True if want to keep the BAM file produced (with mapped and unmapped reads)
+    threads : int, default 1
+        Number of threads to be used
+
+    Returns
+    -------
+    run_successfully : bool
+        Boolean stating if INNUca Assembly_Mapping module ran successfully or not
+    sam_file : str or None
+        If everything went fine, it returns the path to the sam file, otherwise it returns None
+    """
+
     sam_file = os.path.join(outdir, str('alignment.sam'))
 
     # Index reference file
-    run_successfully = indexSequenceBowtie2(referenceFile, threads)
+    run_successfully = indexSequenceBowtie2(reference_file, threads)
 
     if run_successfully:
-        command = ['bowtie2', '-q', '--very-sensitive-local', '--threads', str(threads), '-x', referenceFile, '', '--no-unal', '-S', sam_file]
+        command = ['bowtie2', '-q', '--very-sensitive-local', '--threads', str(threads), '-x', reference_file, '',
+                   '', '-S', sam_file]
         if len(fastq_files) == 1:
-            command[8] = '-U ' + fastq_files[0]
+            command[7] = '-U ' + fastq_files[0]
         else:
-            command[8] = '-1 ' + fastq_files[0] + ' -2 ' + fastq_files[1]
+            command[7] = '-1 ' + fastq_files[0] + ' -2 ' + fastq_files[1]
+
+        if not keep_bam:
+            command[8] = '--no-unal'
+
         run_successfully, stdout, stderr = utils.runCommandPopenCommunicate(command, False, None, True)
 
     if not run_successfully:
@@ -91,9 +120,50 @@ pilon_timer = partial(utils.timer, name='Pilon')
 
 
 @pilon_timer
-def runPilon(jar_path_pilon, assembly, fastq_files, threads, outdir, jarMaxMemory, alignment_file):
-    failing = {}
-    failing['sample'] = False
+def run_pilon(jar_path_pilon, assembly, fastq_files, outdir, jar_max_memory, alignment_file, keep_bam=False, threads=1):
+    """
+    Runs Assembly_Mapping for INNUca and QA/QC the results
+
+    Parameters
+    ----------
+    jar_path_pilon
+    assembly : str
+        Path to the assembly to correct
+    fastq_files : list
+        List of fastq files
+    outdir : str
+        Path to the output directory
+    jar_max_memory : int or 'off'
+        If not 'off' is provided, sets the maximum RAM Gb usage by jar files
+    alignment_file : str or None
+        Path to the BAM file to be used. If None is provided, new alignment reads will be performed
+    keep_bam : bool, default False
+        True if want to keep the BAM file produced (with mapped and unmapped reads)
+    threads : int, default 1
+        Number of threads to be used
+
+    Returns
+    -------
+    run_successfully : bool
+        Boolean stating if INNUca Assembly_Mapping module ran successfully or not
+    pass_qc : None
+        QA/QC not performed
+    time_taken : float
+        Seconds that run_assembly_mapping took to run
+    failing : dict
+        Dictionary with the failing reasons. If sample did not fail, it is only {'sample': False}. If it failed, keys
+        will be the level of failing, and values list of strings
+    assembly_polished : str or None
+        Path to the polished assembly. If something went wrong, None is returned
+    pilon_folder : str
+        Path to Pilon working directory
+    new_bam : bool
+        True if new alignment reads was performed
+    alignment_file : str or None
+        Path to the BAM file used to correct the assembly. If something went wrong, None is returned.
+    """
+
+    failing = {'sample': False}
 
     pilon_folder = os.path.join(outdir, 'pilon', '')
     utils.removeDirectory(pilon_folder)
@@ -105,12 +175,15 @@ def runPilon(jar_path_pilon, assembly, fastq_files, threads, outdir, jarMaxMemor
 
     run_successfully = True
 
+    new_bam = False
     if alignment_file is None:
         # Index assembly using Bowtie2
         run_successfully = indexSequenceBowtie2(assembly_link, threads)
 
         if run_successfully:
-            run_successfully, sam_file = mappingBowtie2(fastq_files, assembly_link, threads, pilon_folder)
+            # mapping_bowtie2(fastq_files, reference_file, outdir, keep_bam=False, threads=1
+            run_successfully, sam_file = mapping_bowtie2(fastq_files=fastq_files, reference_file=assembly_link,
+                                                         outdir=pilon_folder, keep_bam=keep_bam, threads=threads)
 
             if run_successfully:
                 alignment_file = os.path.splitext(sam_file)[0] + '.bam'
@@ -119,22 +192,29 @@ def runPilon(jar_path_pilon, assembly, fastq_files, threads, outdir, jarMaxMemor
                 if run_successfully:
                     os.remove(sam_file)
                     run_successfully = indexAlignment(alignment_file)
+                    new_bam = True
+                else:
+                    alignment_file = None
 
     assembly_polished = None
 
     if run_successfully:
-        run_successfully, assembly_polished = pilon(jar_path_pilon, assembly_link, alignment_file, pilon_folder, jarMaxMemory)
+        run_successfully, assembly_polished = pilon(jar_path_pilon, assembly_link, alignment_file, pilon_folder,
+                                                    jar_max_memory)
 
         if run_successfully:
             parsePilonResult(assembly_polished, outdir)
             shutil.copyfile(assembly_polished, os.path.join(outdir, os.path.basename(assembly_polished)))
             assembly_polished = os.path.join(outdir, os.path.basename(assembly_polished))
+            if keep_bam and new_bam:
+                shutil.copyfile(alignment_file, os.path.join(outdir, os.path.basename(alignment_file)))
+                alignment_file = os.path.join(outdir, os.path.basename(alignment_file))
 
-    if alignment_file is not None and os.path.isfile(alignment_file):
+    if alignment_file is not None and os.path.isfile(str(alignment_file)) and not keep_bam:
         os.remove(alignment_file)
 
     if not run_successfully:
         failing['sample'] = 'Did not run'
         print failing['sample']
 
-    return run_successfully, None, failing, assembly_polished, pilon_folder
+    return run_successfully, None, failing, assembly_polished, pilon_folder, new_bam, alignment_file
